pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stocks-portfolio"
version = "1.0.0"
description = "Portfolio de Stocks"
readme = "README.md"
authors = [{ name = "Fernando Figueroa", email = "me@fena.dev" }]
license = { file = "LICENSE" }
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.2.3",
    "yfinance>=0.2.50",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.1.1",
    "mypy>=1.8.0",
    "commitizen>=3.18.0",
    "pre-commit>=3.6.2",
    "flake8>=7.0.0",
    "isort>=5.13.2",
]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["stocks"]
omit = ["tests/*", "setup.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
